chore: combine AddReview & Review list to reduce read firestore
