--- src/components/AddReview.module.css
@@ -1,22 +1,0 @@
-.rateBar {
-  text-align: center;
-  padding: 10px 0;
-}
-
-.reviewText {
-  width: 100%;
-  border-radius: 10px;
-  border: 1px solid #dfe1e5;
-  margin-bottom: 1rem;
-  padding: 16.5px 14px;
-}
-
-.reviewBtnContainer {
-  width: 100%;
-  text-align: end;
-}
-
-.starIcon {
-  width: 22px;
-  margin: 0 1px;
-}
