+++ src/components/Review.js
@@ -1,22 +1,43 @@
 import { useEffect, useState } from "react";
+import { useNavigate } from "react-router-dom";
 import { ref, update } from "firebase/database";
-import { collection, doc, getDocs, updateDoc } from "firebase/firestore/lite";
+import {
+  collection,
+  doc,
+  addDoc,
+  getDocs,
+  deleteDoc,
+} from "firebase/firestore/lite";
 import database, { firebaseAuth, firestore } from "../service/firebase";
 import {
   Box,
+  Button,
   Paper,
-  Typography,
   Rating,
+  Typography,
+  InputBase,
   Tooltip,
   IconButton,
 } from "@mui/material";
 import { faStar } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import DeleteIcon from "@mui/icons-material/Delete";
-import styles from "./ReviewList.module.css";
+import styles from "./Review.module.css";
 
-function ReviewList({ restaurant }) {
+const labels = {
+  0: "이 식당을 평가해주세요!",
+  1: "Useless",
+  2: "Poor",
+  3: "Ok",
+  4: "Good",
+  5: "Excellent",
+};
+
+function Review({ restaurant }) {
+  const navigate = useNavigate();
   const user = firebaseAuth.currentUser;
+  const [star, setStar] = useState(0);
+  const [review, setReview] = useState("");
   const [reviewList, setReviewList] = useState([]);
   const restaurantRef = ref(database, `restaurants/${restaurant.id}`);
   const reviewsCol = collection(firestore, "reviews");
@@ -25,21 +46,53 @@ function ReviewList({ restaurant }) {
   const starCount = restaurant.starCount;
   const starSum = restaurant.starSum;
   useEffect(() => {
-    async function getReviewList() {
-      const reviewSnapshot = await getDocs(reviewsCol);
-      setReviewList(
-        reviewSnapshot.docs.map((doc) => ({
-          ...doc.data(),
-          id: doc.id,
-        }))
-      );
-    }
     getReviewList();
-  }, [reviewCount]);
-  async function removeReview(reviewId, uid, star) {
-    await updateDoc(doc(firestore, "reviews", reviewId), {
-      visible: false,
+  }, []);
+  async function addReview() {
+    if (!user) {
+      if (window.confirm("로그인 하시겠습니까?"))
+        navigate({ pathname: "/login" });
+      return;
+    }
+    if (restaurant.reviewedUser && restaurant.reviewedUser.includes(user.uid)) {
+      alert("이미 리뷰를 작성하셨습니다 🥺");
+      return;
+    }
+    if (!star) {
+      alert("별점을 매겨주세요 🥺");
+      return;
+    }
+    await addDoc(reviewsCol, {
+      star: star,
+      review: review,
+      uploadTime: new Date(),
+      uid: user.uid,
+      displayName: user.displayName,
+      photoURL: user.photoURL,
+      restaurantId: restaurant.id,
+      visible: true,
     }).then(() => {
+      update(restaurantRef, {
+        reviewCount: reviewCount + 1,
+        starCount: starCount + 1,
+        starSum: starSum + star,
+        reviewedUser: reviewedUser ? [...reviewedUser, user.uid] : [user.uid],
+      });
+    });
+    getReviewList();
+    resetValues();
+  }
+  async function getReviewList() {
+    const reviewSnapshot = await getDocs(reviewsCol);
+    setReviewList(
+      reviewSnapshot.docs.map((doc) => ({
+        ...doc.data(),
+        id: doc.id,
+      }))
+    );
+  }
+  async function removeReview(reviewId, uid, star) {
+    await deleteDoc(doc(firestore, "reviews", reviewId)).then(() => {
       update(restaurantRef, {
         reviewCount: reviewCount - 1,
         starCount: starCount - 1,
@@ -47,6 +100,7 @@ function ReviewList({ restaurant }) {
         reviewedUser: reviewedUser.filter((user) => user !== uid),
       });
     });
+    getReviewList();
   }
   // https://gofnrk.tistory.com/117
   function displayedAt(uploadTime) {
@@ -66,8 +120,46 @@ function ReviewList({ restaurant }) {
     const years = days / 365;
     return `${Math.floor(years)}년 전`;
   }
+  const resetValues = () => {
+    setStar(0);
+    setReview("");
+  };
   return (
     <>
+      <Paper className={styles.section}>
+        <Box className={styles.formRateBar}>
+          <Typography gutterBottom>{labels[star ? star : 0]}</Typography>
+          <Rating
+            value={star}
+            onChange={(event, newValue) => {
+              setStar(newValue);
+            }}
+            icon={
+              <FontAwesomeIcon className={styles.formStarIcon} icon={faStar} />
+            }
+            emptyIcon={
+              <FontAwesomeIcon
+                className={styles.formStarIcon}
+                style={{ opacity: 0.55 }}
+                icon={faStar}
+              />
+            }
+          />
+        </Box>
+        <InputBase
+          className={styles.formReviewText}
+          value={review}
+          onChange={(e) => setReview(e.target.value)}
+          multiline
+          rows={3}
+          placeholder="리뷰를 적어주세요"
+        />
+        <Box className={styles.addBtnContainer}>
+          <Button onClick={addReview} variant="contained" color="secondary">
+            리뷰 추가
+          </Button>
+        </Box>
+      </Paper>
       {reviewList
         .filter(
           (review) => review.visible && review.restaurantId === restaurant.id
@@ -133,4 +225,4 @@ function ReviewList({ restaurant }) {
   );
 }
 
-export default ReviewList;
+export default Review;
