+++ src/components/Review.module.css
@@ -1,3 +1,33 @@
+.section {
+  padding: 1rem;
+  margin-bottom: 24px;
+  border-radius: 1rem;
+  box-shadow: 0 4px 6px -1px rgb(0 0 0/0.1), 0 2px 4px -2px rgb(0 0 0/0.1);
+}
+
+.formRateBar {
+  text-align: center;
+  padding: 10px 0;
+}
+
+.formReviewText {
+  width: 100%;
+  border-radius: 10px;
+  border: 1px solid #dfe1e5;
+  margin-bottom: 1rem;
+  padding: 16.5px 14px;
+}
+
+.addBtnContainer {
+  width: 100%;
+  text-align: end;
+}
+
+.formStarIcon {
+  width: 22px;
+  margin: 0 1px;
+}
+
 .container {
   display: flex;
   padding: 0 1rem;
